Cleanup and make this a python package.
