+++ airhockey/__init__.py
@@ -1,0 +1,3 @@
+from airhockey.airhockey import AirHockeyEnv
+import airhockey.renderers as renderers
+import airhockey.sims as sims
