+++ airhockey/renderers/__init__.py
@@ -1,0 +1,1 @@
+from airhockey.renderers.render import AirHockeyRenderer
