+++ airhockey/sims/__init__.py
@@ -1,0 +1,1 @@
+from airhockey.sims.airhockey_box2d import AirHockeyBox2D
