+++ scripts/create_rollout_figure.py
