+++ scripts/plot_all.py
@@ -75,7 +75,6 @@ def plot_task_performances(base_dir='baseline_models'):
     plt.title('Box2D Task Performance Over Training Steps')
     plt.legend()
     plt.grid()
-    # save for plot in research paper
     # cut off excess space
     plt.savefig('box2d_training_performance.pdf', bbox_inches='tight')
